requirements.txt
Babel==2.7.0
boto3==1.9.215
botocore==1.12.215
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
dialogflow==0.6.0
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
docutils==0.15.2
google-api-core==1.14.2
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.23.0
hvac==0.5.0
idna==2.6
image==1.5.27
jmespath==0.9.4
numpy==1.17.0
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
python-dateutil==2.8.0
pytz==2019.2
PyYAML==3.12
requests==2.18.4
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.22
Wikidata==0.6.1
